add app-drawer UI component
