+++ config/apps/list.js
@@ -17,7 +17,7 @@ export const appList = [
   },
   {
     name: 'HelpCenter',
-    label: 'Help Center',
+    label: 'Help-Center',
     css: 'fa fa-help-circle fs-36 warning--text',
     type: 'route',
     route: '/apps/help-center/',
@@ -30,6 +30,14 @@ export const appList = [
     type: 'url',
     icon: 'https://survey.bitpod.io/favicon.ico',
     href: 'https://dev-survey.bitpod.io/',
+    roles: ['$orguser'],
+  },
+  {
+    name: 'seatmap',
+    label: 'Seat Map',
+    css: 'fa fa-grid-alt fs-36 primary--text',
+    type: 'route',
+    route: '/apps/seatmap/list/seatmaps/seatmaps',
     roles: ['$orguser'],
   },
 ]
